% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{curve_info}
\alias{curve_info}
\title{Metadata for development curves}
\format{
A data frame with 3000 rows and 4 columns:
\describe{
\item{index}{Curve index}
\item{stage}{Development stage; one of L2, L3, L4, L5, L6}
\item{run.index}{Posterior sample index}
\item{colony}{SBW colony membership; one of AB, IPQL, NB, NWT, ON, QC}
\item{s_eps}{Standard deviation of individual multipliers on median dev rate}
}
}
\usage{
curve_info
}
\description{
Metadata for development curves
}
\keyword{datasets}
