% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{curves}
\alias{curves}
\title{Estimated development rates at treatment temperatures}
\format{
A data frame with 27000 rows and 3 columns:
\describe{
\item{index}{Curve index}
\item{temp}{Treatment temperature}
\item{rate}{Estimated Development Rate}
}
}
\usage{
curves
}
\description{
Estimated development rates at treatment temperatures
}
\keyword{datasets}
