% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_days.R
\name{dev_days}
\alias{dev_days}
\title{Predict days to development given a set of daily or hourly temperature data}
\usage{
dev_days(
  weather,
  sbwcolony = "NB",
  period = "hour",
  stage = "Pupa",
  ecdf = FALSE,
  n.post = 100,
  individuals = 100
)
}
\arguments{
\item{weather}{user-defined weather dataset; requires columns:
Temp (temperature in degrees Celsius), Date (YYYY-MM-DD), Hour (integer, 0-23)}

\item{sbwcolony}{colony identifier; one of 'AB', 'IPQL', 'NB', 'NWT', 'ON', 'QC'}

\item{period}{observation interval of weather data; can be one of 'hour' or 'day'}

\item{stage}{larval development stage to be predicted; one of 'L3', 'L4', 'L5',
'L6', 'Pupa' or 'All'}

\item{ecdf}{if TRUE, return empirical CDF (density object) of predicted days}

\item{n.post}{number of posterior samples to simulate from (max 100)}

\item{individuals}{number of individuals in the population per posterior sample}
}
\value{
Matrix or array of Julian dates; dimensions are \code{individuals*n.post*stages}
}
\description{
Predict days to development given a set of daily or hourly temperature data
}
